More verbose error message when a Swagger path is not implemented
If one is not very keenly enlightened in Mojolicious, it is hard
to understand how namespaces and controllers and actions interact.

It is more difficult when such complexity is hidden behind a plugin.

Also it might not be obvious that the value in 'x-mojo-to' is appended
after a namespace and how that might horribly mangle the controller
lookups.

It is very helpful to get a clear error message when such things arise.
Telling clearly what Controller::Subroutine-combo was been looked for and
for which path.

This patch extends the NOT_IMPLEMENTED-error with a clear description
of the error.

Instead of a concise "Not implemented."-error
An error like this is returned:
"OpenAPI path '\/api\/v1\/auth' not implemented in controller
    'PatronStore::Controller::Api::V1::Authenticate->get'."

This will give much more ammunition in solving the outstanding issue
with no need of extensively debugging Mojolicious internals to come up
with a clear error description.