Regression tests to highlight #242
To reproduce:
1. Apply this patch
2. Run:
   $ PERL5LIB=../json-validator/lib/:../mojo/lib:./lib prove t/v2-headers.t
=> FAIL: Tests fail! Weird header handling!

 #   Failed test 'exact match for JSON Pointer "/first"'
 #   at t/v2-headers.t line 50.
 #          got: '42, 24,  44'
 #     expected: '42,24, 44'

 #   Failed test 'exact match for JSON Pointer "/second"'
 #   at t/v2-headers.t line 50.
 #          got: '42'
 #     expected: '42,24, 44'
 # Looks like you failed 2 tests of 32.

The 'first' use case, might be correct, meaning some sanitization of the
collection format takes place. But the fact that the space is added even
on the case that already had one, is suspicious.

The second failure confirms that repeated headers are mangled in a way
they get lost.